added display and users route in controllers folder
added display routes users routes in controllers folder
