+++ backEnd/src/server.js
@@ -9,6 +9,8 @@ const { Client } = require("ssh2");
 const port = 5555;
 const app = express();
 const health=require('./controllers/health');
+const users=require('./controllers/users')
+const display=require('./controllers/display')
 
 app.use(bodyParser.json());
 require("dotenv").config({ path: path.join(__dirname, ".env") });
@@ -18,33 +20,9 @@ require("dotenv").config({ path: path.join(__dirname, ".env") });
 //~~DB OPERATIONS~~
 //================================================================================================================================================================================
 //================================================================================================================================================================================
-app.post(
-  "/users/signup",
-  async ({ body: { firstName, lastName, username, password, email } }, res) => {
-    User.findOne({ username: username }, "username").then((resultant) => {
-      if (resultant) {
-        res.send("Username already exists!");
-      } else {
-        bcrypt.genSalt(10, (err, salt) => {
-          bcrypt.hash(password, salt, (err, hash) => {
-            const newUser = new User({
-              firstName: firstName,
-              lastName: lastName,
-              username: username,
-              hash: hash,
-              email: email,
-              servers: [],
-            });
-            newUser
-              .save()
-              .then(() => res.send("User created."))
-              .catch((err) => res.send(err));
-          });
-        });
-      }
-    });
-  }
-);
+app.use('/users',users)
+
+
 //================================================================================================================================================================================
 //================================================================================================================================================================================
 /*app.post("/users/signin", ({ body: { username, password } }, res) => {
@@ -57,142 +35,21 @@ app.post(
     })
     .catch((err) => res.send(err));    REMOVE THIS COMMENTED CODE ONCE VERIFIED FRONTEND FUNCTIONALITY.
 });*/
-app.post("/users/signin", ({ body: { username, password } }, res) => {
-  User.findOne({ username: username }, "username hash").exec(
-    async (err, result) => {
-      if (result) {
-        const { hash, servers } = result;
-        const authenticationSuccessful = await bcrypt.compare(password, hash);
-        if (authenticationSuccessful) {
-          res.send("Authentication Successful");
-        } else {
-          res.send("Authentication Failed");
-        }
-      } else {
-        res.send("Authentication Failed");
-      }
-    }
-  );
-});
+
 //================================================================================================================================================================================
 //================================================================================================================================================================================
-app.post(
-  "/users/addserver",
-  ({ body: { username, user, serverName, ipAddr, sshKey, password } }, res) => {
-    User.findOne({ username: username }).exec((err, resultant) => {
-      if (err) {
-        throw err;
-        res.send("An error occurred.");
-      } else {
-        if (resultant) {
-          if (resultant.servers.length === 0) {
-            resultant.servers.push({
-              user: user,
-              serverName: serverName,
-              ipAddr: ipAddr,
-              sshKey: sshKey,
-              password: password,
-            });
-            resultant.save();
-            res.send("Added user's first server!");
-          } else {
-            const checkNameAvailable = (server) =>
-              server.serverName === serverName;
-            const nameNotAvailable = resultant.servers
-              .map(checkNameAvailable)
-              .some((el) => el === true);
-            if (nameNotAvailable) {
-              res.send(
-                "Server with that name already exists. Did not add server."
-              );
-            } else {
-              const newServer = {
-                serverName: serverName,
-                ipAddr: ipAddr,
-                sshKey: sshKey,
-                password: password,
-                user: user,
-              };
-              resultant.servers.push(newServer);
-              resultant.save();
-              res.send("Added server.");
-            }
-            //addpasswordencryption for server passwords
-          }
-        } else {
-          res.send("User not found. Failed to add server.");
-        }
-      }
-    });
-  }
-);
+
 //================================================================================================================================================================================
 //================================================================================================================================================================================
-app.post("/users/removeserver", ({ body: { username, serverName } }, res) => {
-  User.findOne({ username: username }).exec((err, resultant) => {
-    if (err) {
-      res.send("An error ocurred");
-    } else {
-      const serverExists = resultant.servers
-        .map((el) => el.serverName === serverName)
-        .some((el) => el === true);
-      if (serverExists) {
-        const serverIndex = resultant.servers
-          .map((el) => el.serverName === serverName)
-          .indexOf(true);
-        resultant.servers.splice(serverIndex, 1);
-        resultant.save();
-        res.send(`Server : ${serverName} was deleted successfully.`);
-      } else {
-        res.send("Server with that name does not exist.");
-      }
-    }
-  });
-});
+
 //================================================================================================================================================================================
 //================================================================================================================================================================================
-app.post("/users/getservers", ({ body: { username, password } }, res) => {
-  User.findOne({ username: username }).exec(async (err, resultant) => {
-    if (err) {
-      res.send("An error ocurred.");
-    } else {
-      if (resultant) {
-        const { hash, servers } = resultant;
-        const authenticationSuccessful = await bcrypt.compare(password, hash);
-        if (authenticationSuccessful) {
-          res.send(servers);
-        } else {
-          res.send("Authentication failed. Did not retrieve servers.");
-        }
-      } else {
-        res.send("Authentication failed. Did not retrieve servers.");
-      }
-    }
-  });
-});
+
+
+
 //================================================================================================================================================================================
 //================================================================================================================================================================================
 app.use('/health',health)
-
-// app.post("/health/setupserver", ({ body: { username, serverName } }, res) => {
-//   User.findOne({ username: username }).exec(async (err, resultant) => {
-//     if (resultant) {
-//       const { servers, hash } = resultant;
-//       const authenticationSuccessful = true; // await bcrypt.compare(password, hash);
-//       if (authenticationSuccessful) {
-//         const serverIndex = servers
-//           .map((el, index) => el.serverName === serverName)
-//           .indexOf(true);
-//         const { user, password, ipAddr } = servers[serverIndex];
-//         const command = await sshInit(user, password, ipAddr, res);
-//       } else {
-//         res.send("Authentication failed.");
-//       }
-//     } else {
-//       res.send("User not found.");
-//     }
-//   });
-// });
 //================================================================================================================================================================================
 //================================================================================================================================================================================
 //~~END OF DB OPERATIONS
@@ -243,72 +100,8 @@ async function sshInit(username, password, host, res) {
 //================================================================================================================================================================================
 //~Connecting to Health Server Application
 //================================================================================================================================================================================
-app.post(
-  "/display",
-  ({ body: { username, password, user, serverName, details } }, res) => {
-    User.findOne({ username: username }, "hash servers").exec(
-      async (err, result) => {
-        if (err) {
-          res.json({ message: err });
-        } else {
-          if (result) {
-            const name = serverName;
-            const { hash, servers } = result;
-            let ip = "";
-            let serverUsername = "";
-            if (servers.length === 0) {
-              res.json({ message: "This user does not have any servers" });
-            }
-            for (let i = 0; i < servers.length; i++) {
-              const { serverName, ipAddr, user } = servers[i];
-              if (serverName === name) {
-                ip = ipAddr;
-                serverUsername = user;
-                break;
-              }
-            }
-            if (ip === "" || serverUsername === "") {
-              res.json({
-                message:
-                  "server with that name not found or server username is invalid.",
-              });
-            }
-            const checkAuthentication = await bcrypt.compare(password, hash);
-            if (checkAuthentication) {
-              request.post(
-                {
-                  url: `http://${ip}:4400/Display`, //shouldn't this be the IP of the user's server.
-                  json: {
-                    Username: user,
-                    Servername: serverName,
-                    Details: details, //again wtf is details.
-                  },
-                  headers: {
-                    "Content-type": "application/json",
-                  },
-                },
-                (err, response) => {
-                  //callback to be executed once request has been sent.
-                  if (err) {
-                    res.json({ message: err });
-                  } else {
-                    res.send(response.body);
-                  }
-                }
-              );
-            } else {
-              res.json({
-                message: "Data retrieval failed. Invalid credentials.",
-              });
-            }
-          } else {
-            res.json({ message: "User does not exist." });
-          }
-        }
-      }
-    );
-  }
-);
+app.use('/display',display)
+
 //================================================================================================================================================================================
 //================================================================================================================================================================================
 //================================================================================================================================================================================
